add new functions, fixed Makefile

+++ ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: akolomii <akolomii@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/01 14:17:24 by akolomii          #+#    #+#             */
-/*   Updated: 2025/10/02 13:11:26 by anth             ###   ########.fr       */
+/*   Updated: 2025/10/02 13:11:26 by akolomii         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h";
+#include "libft.h"
 
-void	*ft_memset(void *s, int c, size_t n)
+void	*ft_memset(void *s, int c, int n)
 {
-	size_t			i;
+	int				i;
 	unsigned char	*p;
 
 	i = 0;
